Paso 23 completo

+++ API/facturacion/src/test/java/org/una/tienda/facturacion/services/ClienteServiceImplementationTest.java
@@ -9,12 +9,16 @@ import java.util.Optional;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Assertions;
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.fail;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.una.tienda.facturacion.dto.ClienteDTO;
+import org.una.tienda.facturacion.exceptions.ClienteConDireccionException;
+import org.una.tienda.facturacion.exceptions.ClienteConEmailException;
+import org.una.tienda.facturacion.exceptions.ClienteConTelefonoException;
 
 /**
  *
@@ -22,12 +26,18 @@ import org.una.tienda.facturacion.dto.ClienteDTO;
  */
 @SpringBootTest
 public class ClienteServiceImplementationTest {
-    
+
     @Autowired
     private IClienteService clienteService;
 
     ClienteDTO clienteEjemplo;
-    
+
+    ClienteDTO clientePrueba;
+
+    ClienteDTO clientePrueba2;
+
+    ClienteDTO clientePrueba3;
+
     @BeforeEach
     public void setup() {
         clienteEjemplo = new ClienteDTO() {
@@ -35,13 +45,45 @@ public class ClienteServiceImplementationTest {
                 setDireccion("Pérez");
                 setEmail("@@@");
                 setNombre("Luis");
-                setTelefono("123"); 
+                setTelefono("123");
             }
         };
     }
 
+    private void initDataClienteConTelefono() {
+        clientePrueba = new ClienteDTO() {
+            {
+                setDireccion("Pérez");
+                setEmail("@@@");
+                setNombre("Luis");
+            }
+        };
+
+    }
+
+    private void initDataClienteConDirección() {
+        clientePrueba2 = new ClienteDTO() {
+            {
+                setTelefono("123");
+                setEmail("@@@");
+                setNombre("Luis");
+            }
+        };
+    }
+
+    private void initDataClienteConEmail() {
+        clientePrueba3 = new ClienteDTO() {
+            {
+                setDireccion("Pérez");
+                setNombre("Luis");
+                setTelefono("123");
+            }
+        };
+
+    }
+
     @Test
-    public void sePuedeCrearUnClienteCorrectamente() {
+    public void sePuedeCrearUnClienteCorrectamente() throws ClienteConTelefonoException, ClienteConEmailException, ClienteConDireccionException {
 
         clienteEjemplo = clienteService.create(clienteEjemplo);
 
@@ -72,7 +114,43 @@ public class ClienteServiceImplementationTest {
     }
 
     @Test
-    public void sePuedeEliminarUnClienteCorrectamente() {
+    public void seEvitaCrearUnClienteSinTelefono() {
+
+        initDataClienteConTelefono();
+
+        assertThrows(ClienteConTelefonoException.class,
+                () -> {
+                    clienteService.create(clientePrueba);
+                }
+        );
+    }
+
+    @Test
+    public void seEvitaCrearUnClienteSinDireccion() {
+
+        initDataClienteConDirección();
+
+        assertThrows(ClienteConDireccionException.class,
+                () -> {
+                    clienteService.create(clientePrueba2);
+                }
+        );
+    }
+
+    @Test
+    public void seEvitaCrearUnClienteSinEmail() {
+
+        initDataClienteConEmail();
+
+        assertThrows(ClienteConEmailException.class,
+                () -> {
+                    clienteService.create(clientePrueba3);
+                }
+        );
+    }
+
+    @Test
+    public void sePuedeEliminarUnClienteCorrectamente() throws ClienteConTelefonoException, ClienteConEmailException, ClienteConDireccionException {
 
         clienteEjemplo = clienteService.create(clienteEjemplo);
 
@@ -91,7 +169,9 @@ public class ClienteServiceImplementationTest {
     @AfterEach
     public void tearDown() {
         if (clienteEjemplo != null) {
-            clienteService.delete(clienteEjemplo.getIdclientes());
+            if (clienteEjemplo.getIdclientes() != null) {
+                clienteService.delete(clienteEjemplo.getIdclientes());
+            }
             clienteEjemplo = null;
         }
